Lost again sound references and added some stuff

+++ Assets/Scripts/Spawn.cs
@@ -14,7 +14,7 @@ public class Spawn : NetworkBehaviour
 
     public LayerMask vehicleLayer;
 
-    public int maxTimeSpawn;
+    public float maxTimeSpawn;
     public int vehicleCount = 0;
     public int maxVehicles;
     public float vehicleSpeed = 75f;
